backend/services/llm_gateway/requirements.txt
# Core web framework
fastapi
uvicorn[standard]

# For environment variable management
pydantic-settings

# Google Gemini API client library
google-generativeai
